correct harvest and plant

+++ CryptomonServer/Dtos/AccountDto.cs
@@ -6,7 +6,7 @@
         public string Address { get; set; } // Unique account address (the Ethereum account)
         public string Username { get; set; } // The user name
         public string RecoveryEmail { get; set; } // The user Email
-        public long CoinBalance { get; set; }
-        public long MinaBalance { get; set; }
+        public decimal CoinBalance { get; set; }
+        public decimal MinaBalance { get; set; }
     }
 }
